+++ android/src/main/java/com/appsflyer/reactnative/RNAppsFlyerModule.java
@@ -727,4 +727,20 @@ public class RNAppsFlyerModule extends ReactContextBaseJavaModule {
     public void disableAdvertisingIdentifier(Boolean isDisabled) {
         AppsFlyerLib.getInstance().setDisableAdvertisingIdentifiers(isDisabled);
     }
+
+    @ReactMethod
+    public void setPartnerData(String partnerId, ReadableMap partnerData) {
+        Map partnerDataAsMap = RNUtil.toMap(partnerData);
+        AppsFlyerLib.getInstance().setPartnerData(partnerId, partnerDataAsMap);
+    }
+
+    @ReactMethod
+    public void addListener(String eventName) {
+      // Keep: Required for RN built in Event Emitter Calls.
+    }
+  
+    @ReactMethod
+    public void removeListeners(Integer count) {
+      // Keep: Required for RN built in Event Emitter Calls.
+    }
 }